pyproject.toml
[project]
name = "2025-26-dl-assignments"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "datasets>=4.8.5",
    "optuna-dashboard>=0.20.0",
    "pandas>=3.0.1",
    "seaborn>=0.13.2",
    "sklearn-preprocessing>=0.1.0",
    "standardscaler>=0.5",
    "statsmodels>=0.14.6",
    "tensorboard>=2.20.0",
    "torch>=2.10.0",
    "torchaudio>=2.10.0",
    "torchtext>=0.18.0",
    "torchvision>=0.25.0",
]
